Renamed HandView to ViewModel, Fixed 1.17 empty hand ViewModel bug
