+++ ares-fabric-1.16/src/main/java/dev/tigr/ares/fabric/AresMod.java
@@ -156,7 +156,7 @@ public class AresMod extends Ares {
                 ESP.class,
                 ExtraTab.class,
                 FullBright.class,
-                HandView.class,
+                ViewModel.class,
                 HoleESP.class,
                 MapTooltips.class,
                 MobOwner.class,
