+++ ares-fabric-1.16/src/main/resources/ares.mixins.json
@@ -17,7 +17,6 @@
     "accessors.EntityAccessor",
     "accessors.EntityVelocityUpdateS2CPacketAccessor",
     "accessors.ExplosionS2CPacketAccessor",
-    "accessors.HeldItemRendererAccessor",
     "accessors.KeyBindingAccessor",
     "accessors.MatrixStackAccessor",
     "accessors.MinecraftClientAccessor",
