--- ares-fabric/src/main/java/dev/tigr/ares/fabric/mixininterface/IHeldItemRenderer.java
@@ -1,14 +1,0 @@
-package dev.tigr.ares.fabric.mixininterface;
-
-import net.minecraft.client.network.AbstractClientPlayerEntity;
-import net.minecraft.client.render.VertexConsumerProvider;
-import net.minecraft.client.util.math.MatrixStack;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.Hand;
-
-public interface IHeldItemRenderer {
-    void doRenderFirstPersonItem(
-            AbstractClientPlayerEntity player, float tickDelta, float pitch, Hand hand, float swingProgress,
-            ItemStack item, float equipProgress, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light
-    );
-}
