+++ ares-forge/src/main/java/dev/tigr/ares/forge/AresMod.java
@@ -173,7 +173,7 @@ public class AresMod extends Ares {
                 ESP.class,
                 ExtraTab.class,
                 FullBright.class,
-                HandView.class,
+                ViewModel.class,
                 HoleESP.class,
                 MapTooltips.class,
                 MobOwner.class,
