--- ares-forge/src/main/java/dev/tigr/ares/forge/mixininterface/IItemRenderer.java
@@ -1,9 +1,0 @@
-package dev.tigr.ares.forge.mixininterface;
-
-public interface IItemRenderer {
-    void doRotateArroundXAndY(float angle, float angleY);
-
-    void doSetLightMap();
-
-    void doRotateArm(float p_187458_1_);
-}
